requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=hc377ac9_0
aiohttp=3.8.3=py310h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
asgiref=3.5.2=py310hca03da5_0
async-timeout=4.0.2=py310hca03da5_0
attrs=22.1.0=py310hca03da5_0
blas=2.117=openblas
blas-devel=3.9.0=17_osxarm64_openblas
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.0=h80987f9_0
ca-certificates=2023.05.30=hca03da5_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=h302bd0f_4
certifi=2023.5.7=py310hca03da5_0
cffi=1.15.1=py310h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py310h834c97f_2
cycler=0.11.0=pypi_0
django=4.1=py310hca03da5_0
django-environ=0.8.0=pyhd8ed1ab_0
djangorestframework=3.14.0=pyhd8ed1ab_0
dlib=19.24.2=pypi_0
eigen=3.3.7=h525c30c_1
expat=2.4.9=hc377ac9_0
face-recognition=1.3.0=pypi_0
face-recognition-models=0.3.0=pypi_0
facenet-pytorch=2.5.3=pypi_0
filelock=3.12.2=pypi_0
fontconfig=2.14.1=hee714a5_2
fonttools=4.40.0=pypi_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py310h80987f9_0
gettext=0.21.0=h13f89a0_1
giflib=5.2.1=h80987f9_3
glib=2.69.1=h514c7bf_2
google-api-core=2.11.1=pyhd8ed1ab_0
google-api-core-grpc=2.11.1=hd8ed1ab_0
google-auth=2.21.0=pyh1a96a4e_0
google-cloud-vision=3.4.3=pyhd8ed1ab_0
googleapis-common-protos=1.56.4=py310hca03da5_0
graphite2=1.3.14=hc377ac9_1
grpc-cpp=1.48.2=hc60591f_1
grpcio=1.48.2=py310hc60591f_1
grpcio-status=1.41.1=pyhd3eb1b0_0
gst-plugins-base=1.14.1=h313beb8_1
gstreamer=1.14.1=h80987f9_1
harfbuzz=4.3.0=he9eebac_1
hdf5=1.12.1=h05c076b_3
icu=68.1=hc377ac9_0
idna=3.4=py310hca03da5_0
jinja2=3.1.2=pypi_0
jpeg=9e=h80987f9_1
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libblas=3.9.0=17_osxarm64_openblas
libcblas=3.9.0=17_osxarm64_openblas
libclang=12.0.0=default_hc321e17_4
libcurl=8.1.1=h3e2b118_1
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=12_2_0_hd922786_31
libgfortran5=12.2.0=h0eea778_31
libiconv=1.16=h1a28f6b_2
liblapack=3.9.0=17_osxarm64_openblas
liblapacke=3.9.0=17_osxarm64_openblas
libllvm12=12.0.0=h12f7ac0_4
libnghttp2=1.52.0=h62f6fdd_1
libopenblas=0.3.23=openmp_hc731615_0
libpng=1.6.39=h80987f9_0
libpq=12.15=h02f6b3c_1
libprotobuf=3.20.3=h514c7bf_0
libssh2=1.10.0=h02f6b3c_2
libtiff=4.5.0=h313beb8_2
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
libxml2=2.10.3=h372ba2a_0
libxslt=1.1.37=habca612_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.2=py310h1a28f6b_0
ncurses=6.4=h313beb8_0
networkx=3.1=pypi_0
nspr=4.35=h313beb8_0
nss=3.89.1=h313beb8_0
numpy=1.25.0=py310h3b2db8e_0
numpy-base=1.25.0=py310ha9811e2_0
openblas=0.3.23=openmp_hf78f355_0
opencv=4.6.0=py310he2359d5_2
openssl=3.0.9=h1a28f6b_0
packaging=23.1=pypi_0
pcre=8.45=hc377ac9_0
pillow=10.0.0=pypi_0
pip=23.1.2=py310hca03da5_0
pixman=0.40.0=h1a28f6b_0
proto-plus=1.22.1=py310hca03da5_0
protobuf=3.20.3=py310h313beb8_0
psycopg2=2.9.3=py310h80987f9_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py310hca03da5_0
pyparsing=3.1.0=pypi_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.11=hb885b13_3
python-dateutil=2.8.2=pypi_0
pytz=2022.7=py310hca03da5_0
pyu2f=0.1.5=pyhd8ed1ab_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_4
qtwebkit=5.212=h19f419d_5
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
requests=2.29.0=py310hca03da5_0
rsa=4.7.2=pyhd3eb1b0_1
setuptools=67.8.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h80987f9_0
sqlparse=0.4.3=py310hca03da5_0
sympy=1.12=pypi_0
tk=8.6.12=hb8d0fd4_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py310hca03da5_0
wheel=0.38.4=py310hca03da5_0
xz=5.4.2=h80987f9_0
yarl=1.8.1=py310h1a28f6b_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0


# option plugin
black